+++ src/main/java/org/smoothbuild/lang/expr/BoundValueExpression.java
@@ -1,20 +1,25 @@
 package org.smoothbuild.lang.expr;
 
 import static java.util.Arrays.asList;
-import static org.smoothbuild.task.base.Evaluator.valueEvaluator;
 
 import org.smoothbuild.db.values.ValuesDb;
+import org.smoothbuild.lang.function.base.Name;
 import org.smoothbuild.lang.function.base.Scope;
 import org.smoothbuild.lang.message.Location;
 import org.smoothbuild.lang.type.Type;
 import org.smoothbuild.task.base.Evaluator;
 
-public class DefaultValueExpression extends Expression {
-  public DefaultValueExpression(Type type, Location location) {
+public class BoundValueExpression extends Expression {
+  private final Name name;
+
+  public BoundValueExpression(Type type, Name name, Location location) {
     super(type, asList(), location);
+    this.name = name;
   }
 
   public Evaluator createEvaluator(ValuesDb valuesDb, Scope<Evaluator> scope) {
-    return valueEvaluator(type().defaultValue(valuesDb), location());
+    // TODO this would cause returned evaluator.execute to be called twice
+    // and print twice to console, second time as cached
+    return scope.get(name);
   }
 }