+++ src/main/java/org/smoothbuild/parse/FindSemanticErrors.java
@@ -1,6 +1,7 @@
 package org.smoothbuild.parse;
 
 import static java.util.stream.Collectors.toSet;
+import static org.smoothbuild.lang.function.base.Scope.scope;
 import static org.smoothbuild.util.Lists.map;
 import static org.smoothbuild.util.StringUnescaper.unescaped;
 
@@ -11,12 +12,14 @@ import java.util.Set;
 
 import org.smoothbuild.lang.function.Functions;
 import org.smoothbuild.lang.function.base.Name;
+import org.smoothbuild.lang.function.base.Scope;
 import org.smoothbuild.parse.ast.ArgNode;
 import org.smoothbuild.parse.ast.ArrayTypeNode;
 import org.smoothbuild.parse.ast.Ast;
 import org.smoothbuild.parse.ast.CallNode;
 import org.smoothbuild.parse.ast.FuncNode;
 import org.smoothbuild.parse.ast.ParamNode;
+import org.smoothbuild.parse.ast.RefNode;
 import org.smoothbuild.parse.ast.StringNode;
 import org.smoothbuild.parse.ast.TypeNode;
 import org.smoothbuild.util.UnescapingFailedException;
@@ -28,6 +31,7 @@ public class FindSemanticErrors {
     List<ParseError> errors = new ArrayList<>();
     unescapeStrings(errors, ast);
     overridenBuiltinFunctions(errors, functions, ast);
+    undefinedReferences(errors, ast);
     duplicateFunctions(errors, functions, ast);
     undefinedFunctions(errors, functions, ast);
     duplicateParamNames(errors, ast);
@@ -59,6 +63,33 @@ public class FindSemanticErrors {
           errors.add(new ParseError(function, "Function '" + function.name()
               + "' cannot override builtin function with the same name."));
         }
+      }
+    }.visitAst(ast);
+  }
+
+  private static void undefinedReferences(List<ParseError> errors, Ast ast) {
+    new AstVisitor() {
+      Scope<Name> scope = null;
+
+      public void visitRef(RefNode ref) {
+        super.visitRef(ref);
+        if (!scope.contains(ref.name())) {
+          errors.add(new ParseError(ref.location(), "Unknown parameter '" + ref.name() + "'."));
+        }
+      }
+
+      public void visitFunction(FuncNode func) {
+        scope = scope();
+        func
+            .params()
+            .stream()
+            .forEach(p -> {
+              if (!scope.contains(p.name())) {
+                scope.add(p.name(), null);
+              }
+            });
+        super.visitFunction(func);
+        scope = null;
       }
     }.visitAst(ast);
   }